my-fastapi-project/requirements.txt
fastapi
uvicorn
python-jose
passlib
asyncpg
httpx
python-dotenv
bcrypt